Authraxai-streamlit/requirements.txt
python-dotenv
langchain
openai
pinecone-client
unstructured
tabulate
pdf2image
streamlit
streamlit-option-menu
htmltemplate
tiktoken
unstructured[pdf]
requests
pytesseract
transformers
PyPDF2
uuid
openai
json
txtai